requirements_wenku_to_pdf.txt
Flask
playwright
Pillow
img2pdf
